fix: Unstable search behaviour
According to the GitHub API specification,
The structure of a query consists of one or more KEYWORDs and one
one more QUALIFIERs, as follows.

```
KEYWORD_1 KEYWORD_2 QUALIFIER_1 QUALIFIER_2
```

The character limit of query does not apply to QUALIFIERs, so the
part excluding the QUALIFIERs must be validated.

See-also: https://docs.github.com/en/rest/search/search?apiVersion=2022-11-28#limitations-on-query-length